// projects/formula-python-editor/src/lib/interfaces/acorn/acorn-type.ts
export enum AcornType {
  Program = 'Program',
  Literal = 'Literal',
  CallExpression = 'CallExpression',
  BinaryExpression = 'BinaryExpression',
  ExpressionStatement = 'ExpressionStatement',
  Identifier = 'Identifier',
  ConditionalExpression = 'ConditionalExpression',
  ReturnStatement = 'ReturnStatement',
  UnaryExpression = 'UnaryExpression',
  BlockStatement = 'BlockStatement',
  FunctionDeclaration = 'FunctionDeclaration',
  IfStatement = 'IfStatement',
  LogicalExpression = 'LogicalExpression',
}